requirements.txt
langchain
langchain-openai
langchain-community
sqlmodel
python-dotenv
sqlalchemy
pydantic
fastapi
uvicorn
python-multipart
pytest
httpx
pytest-asyncio
pre-commit
